requirements.txt
python-dotenv
ipykernel
langchain
langchain_core
langchain_community
langchain-text-splitters
langserve
langchain-openai
langchain_huggingface
langchain_groq
langchain-chroma
pypdf
pymupdf
beautifulsoup4
arxiv
lxml
chromadb
sentence_transformers
faiss-cpu
streamlit
fastapi
uvicorn
sse_starlette
arxiv
wikipedia
duckduckgo-search
validators==0.28.1
youtube_transcript_api
unstructured
huggingface_hub
requests
crewai
crewai_tools
neo4j
numexpr
boto3
awscli